Implement a HistogramGenerator
It generates a HistogramBuffer based on a System. It runs on a separate
thread to not interrupt the UI.

The Generator consists of two parts: the Blender and Gatherer.
Blender "blends" the particles according to the rules specified by the
System. Gatherer "gathers" the particles from the Blender and builds a
histogram out of them. The two parts are implemented as tasks
coordinated by the Generator using Qt's signals and event loop.

Implementations are not final — neither performance or feature
completeness were the goal.