Don't scale the image down when modifying weights on a single Formula
It was pretty annoying that if I had only one formula in a blend,
I couldn't really play with the weight, because it also scaled the
result down. To fix that, I've changed the toSource() function to
automatically generate an AST with a linear formula in the mix.

It's actually pretty cool that you can just do that by generating a
different AST. Who knows what I'll do in the future :P